requirements.txt
streamlit
gspread
oauth2client
pyzbar
pillow
numpy
opencv-python-headless
